Change Temperature to nullable float in AgentConfiguration
Updated the Temperature property in AgentConfiguration from double to nullable float to better align with model requirements. Adjusted default value to null, as some models do not support this parameter, and removed unnecessary casting in AgentFactory.

+++ src/AgentFramework.Core/Agents/AgentFactory.cs
@@ -58,7 +58,7 @@ namespace AgentFramework.Core.Agents
                 {
                     Instructions = agentSettings.Instructions,
                     ModelId = agentSettings.Model,
-                    Temperature = (float)(agentSettings.Temperature),
+                    Temperature = agentSettings.Temperature,
                     ResponseFormat = agentSettings.GetResponseFormat(),
                     ToolMode = agentSettings.GetToolMode(),
                     Tools = new AgentToolFactory().GetTools(agentSettings.ToolList),
